Changed default scan mode in A1 launch files to allow default launch to work
The current default param for the A1 launch files, defaults to 'Sensitivity'. This scan mode does not work on the older A1 lidar units. Unfortunately when it fails it still keeps the node running but data is not published from the A1. To correct this behaviour, the launch and view files for the A1 lidar are updated to use scan_mode 'Standard' which has been to tested to work successfully.